feat: Implement infinite scrolling for character list
The code changes include implementing infinite scrolling for the character list in the `app/characters/page.tsx` file. The `react-infinite-scroller` package is added to the `package.json` dependencies, and the `loadFunc` function is created to fetch more characters when the user scrolls to the bottom of the page.

Recent user commits:
- chore: Add loading spinner and confetti component
- feat: Add confetti component and loading spinner
- Adding some css imports and image to main page and new font
- Implementing apollo client for consuming graphql
- some refactorings
- Creating footer for app

Recent repository commits:
- init flowbite
- Initial commit from Create Next App